Regex API refinement
Regex evaluations now return a result, to accommodate the possibility
of the VM exceeding resource bounds.

In addition, this patch adds a captures_from_pos method, which starts
scanning from a given position. This requires a fair amount of
mechanism, including compiling a second inner regex in the cases where
the pattern has 1-character look-behind.