Add flags to regex
Support for flags. A fairly major change, as we had to add new parse
nodes and VM insns to support the new options. In addition, parsing is
now done without regexes, because dealing with optional whitespace and
comments in a regex would be painful (though not at all impossible).

Also, some minor fixes, making things more robust, and more tests, as
well as a bit more writeup in the README.