Add atomic groups
Adds atomic groups, both the (?>foo) group syntax and the a++, a*+,
a?+ possessive syntax.

Note: save state for the begin/end atomic group operation is now saved
on an explicit stack, rather than the reserved slots as for other
similar state (repeat count, lookahead). We'll need that for
(potentially recursive) group calls.

Some optimizations are still possible (including getting rid of the
BeginAtomic and EndAtomic vm instructions when the inner regexp
doesn't backtrack), but nothing too important.