Reject invalid backreferences
Makes sure that any backreference references a group that has already
been closed. We needed to change the signature of the analyze function
to return a result to indicate that it could fail.

Also adds tests for the failing cases and a couple of other invalid
backrefence constructs.

Fixes #2 and #3