Fix anchoring of delegate regular expressions
Before, e.g. a lookahead like `(?=x|a)` would be wrong. The delegate RE
for the `x|a` part inside was compiled as `^x|a`. This means only the
`x` was anchored, but not `a`.

Fix it by adding a non-capturing group so that the delegate is
`^(?:x|a)`, which results in correct anchoring in this case (and
possibly others).