Add "graph" and "trace-inner" commands to toy
The graph command prints a graphviz file to render a VM program to a
nice visual representation. That makes it easier to follow than text
because the possible flows are visible.

Also changed the "trace" command to trace the full program, and make the
old trace "trace-inner". Not sure about the names, but this should
reduce confusion because the full trace can be compared to run/graph.