Fix negative lookaround matching
Finding pattern `(?!a(?=b)|x)` in the text `ab` produced the incorrect
result (0, 0), instead of (1, 1).

The reason for that was that the implementation with the DoubleFail
instruction would only pop 1 state before failing. But in a case like
the above, there is more than 1 state that needs to be popped because of
the alternative.